Fix dark mode: replace hardcoded colors with semantic tokens, reduce duplication
- Replace all hardcoded light-mode Tailwind classes (bg-white, bg-neutral-100/200,
  text-neutral-400/500/600, border-neutral-200/300) with semantic tokens
  (bg-background, bg-card, bg-muted, text-muted-foreground, border-border)
  across 21 components including sidebar, navigation, kanban, calendar, buttons,
  avatars, and layouts
- Fix DottedSeparator hardcoded #D4D4D8 to use hsl(var(--border))
- Add dark: variants to colored icon pill backgrounds in projects sidebar
- Fix button variants (secondary, muted, ghost, disabled) to use semantic tokens
- Fix ghost button border-tranparent typo causing visible borders
- Replace bg-neutral-300 dot separators with bg-border in tasks-list,
  kanban-card, and event-card
- Extract DashboardCard component shared by members-list, projects-list,
  tasks-list to eliminate repeated bg-card wrapper pattern
- Extract TaskFormFields component from create/edit task forms (452 duplicate
  lines → single canonical component); also fixes field resets always firing
  on project change, and decouple from onProjectChange callback
- Fix nested button-inside-Link in SubItem (projects.tsx) for keyboard/a11y
