+++ src/app/(standalone)/layout.tsx
@@ -8,7 +8,7 @@ interface StandaloneLayoutProps {
 
 const StandaloneLayout = ({ children }: StandaloneLayoutProps) => {
 	return (
-		<main className="bg-neutral-100 min-h-screen">
+		<main className="bg-muted min-h-screen">
 			<div className="mx-auto max-w-screen-2xl p-4">
 				<nav className="flex justify-between items-center h-[73px]">
 					<Link href="/">
