+++ src/components/navigation.tsx
@@ -52,12 +52,12 @@ export const Navigation = () => {
 					<Link key={route.href} href={fullHref}>
 						<div
 							className={cn(
-								"flex items-center gap-2.5 p-2.5 rounded-md font-medium hover:text-primary transition text-neutral-500",
+								"flex items-center gap-2.5 p-2.5 rounded-md font-medium hover:text-primary transition text-muted-foreground",
 								isActive &&
-									"bg-white shadow-sm hover:opacity-100 text-primary"
+									"bg-background shadow-sm hover:opacity-100 text-primary"
 							)}
 						>
-							<Icon className="size-5 text-neutral-500" />
+							<Icon className="size-5 text-muted-foreground" />
 							{route.label}
 						</div>
 					</Link>
