+++ src/components/sidebar.tsx
@@ -7,7 +7,7 @@ import { Projects } from "./projects";
 
 export const Sidebar = () => {
 	return (
-		<aside className="h-full bg-neutral-100 p-4 w-full">
+		<aside className="h-full bg-muted p-4 w-full">
 			<Link href="/">
 					<Image
 						src="/logo.svg"
