+++ src/components/ui/button.tsx
@@ -5,7 +5,7 @@ import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
 
 const buttonVariants = cva(
-	"inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-semibold transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:bg-neutral-100 disabled:from-neutral-100 disabled:to-neutral-100 disabled:text-neutral-300 border-neutral-200 border shadow-sm",
+	"inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-semibold transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:bg-muted disabled:from-muted disabled:to-muted disabled:text-muted-foreground/40 border-border border shadow-sm",
 	{
 		variants: {
 			variant: {
@@ -15,9 +15,9 @@ const buttonVariants = cva(
 					"bg-gradient-to-b from-amber-600 to-amber-700 text-destructive-foreground hover:from-amber-700 hover:to-amber-700",
 				outline:
 					"border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground",
-				secondary: "bg-white text-black hover:bg-neutral-100",
-				ghost: "border-tranparent shadow-none hover:bg-accent hover:text-accent-foreground",
-				muted: "bg-neutral-200 text-neutral-600 hover:bg-neutral-200/80",
+				secondary: "bg-card text-foreground hover:bg-muted",
+				ghost: "border-transparent shadow-none hover:bg-accent hover:text-accent-foreground",
+				muted: "bg-muted text-muted-foreground hover:bg-muted/80",
 				teritary:
 					"bg-blue-100 text-blue-600 border-transparent hover:bg-blue-200 shadow-none",
 			},
