+++ src/components/workspace-switcher.tsx
@@ -25,14 +25,14 @@ export const WorkspaceSwitcher = () => {
 	return (
 		<div className="flex flex-col gap-y-2">
 			<div className="flex items-center justify-between">
-				<p className="text-xs uppercase text-neutral-500">Workspaces</p>
+				<p className="text-xs uppercase text-muted-foreground">Workspaces</p>
 				<RiAddCircleFill
 					onClick={open}
-					className="size-5 text-neutral-500 cursor-pointer hover:opacity-75 transition"
+					className="size-5 text-muted-foreground cursor-pointer hover:opacity-75 transition"
 				/>
 			</div>
 			<Select onValueChange={onSelect} value={workspaceId}>
-				<SelectTrigger className="w-full bg-neutral-200 font-medium p-1">
+				<SelectTrigger className="w-full bg-muted font-medium p-1">
 					<SelectValue placeholder="No workspace selected" />
 					<SelectContent>
 						{workspaces?.documents.map((workspace) => (
