+++ src/features/projects/components/edit-project-form.tsx
@@ -155,7 +155,7 @@ export const EditProjectForm = ({
 												) : (
 													<Avatar className="size-[72px]">
 														<AvatarFallback>
-															<ImageIcon className="size-[36px] text-neutral-400" />
+															<ImageIcon className="size-[36px] text-muted-foreground" />
 														</AvatarFallback>
 													</Avatar>
 												)}
