+++ src/features/tasks/components/event-card.tsx
@@ -47,14 +47,14 @@ export const EventCard = ({
 			<div
 				onClick={onClick}
 				className={cn(
-					"p-1.5 text-xs bg-white text-primary border rounded-md border-l-4 flex flex-col gap-y-1.5 cursor-pointer hover:opacity-75 transition",
+					"p-1.5 text-xs bg-card text-primary border rounded-md border-l-4 flex flex-col gap-y-1.5 cursor-pointer hover:opacity-75 transition",
 					statusColorMap[status]
 				)}
 			>
 				<p>{title}</p>
 				<div className="flex items-center gap-x-1">
 					<MemberAvatar name={assignee?.name || "Unknown"} />
-					<div className="size-1 rounded-full bg-neutral-300" />
+					<div className="size-1 rounded-full bg-border" />
 					<ProjectAvatar
 						name={project?.name || "Unknown"}
 						imageUrl={project?.imageUrl}
