+++ src/features/tasks/components/kanban-card.tsx
@@ -14,11 +14,11 @@ interface KanbanCardProps {
 
 export const KanbanCard = ({ task }: KanbanCardProps) => {
 	return (
-		<div className="bg-white p-2.5 mb-1.5 rounded shadow-sm space-y-3">
+		<div className="bg-card p-2.5 mb-1.5 rounded shadow-sm space-y-3">
 			<div className="flex items-start justify-between gap-x-2">
 				<p className="text-sm line-clamp-2">{task.name}</p>
 				<TaskActions id={task.$id} projectId={task.projectId} issueType={task.issueType}>
-					<MoreHorizontalIcon className="size-[18px] stroke-1 shrink-0 text-neutral-700 hover:opacity-75 transition" />
+					<MoreHorizontalIcon className="size-[18px] stroke-1 shrink-0 text-muted-foreground hover:opacity-75 transition" />
 				</TaskActions>
 			</div>
 			{(task.priority || task.issueType) && (
@@ -41,7 +41,7 @@ export const KanbanCard = ({ task }: KanbanCardProps) => {
 					name={task.assignee?.name || "Unknown"}
 					fallbackClassName="text-[10px]"
 				/>
-				<div className="size-1 rounded-full bg-neutral-300" />
+				<div className="size-1 rounded-full bg-border" />
 				<TaskDate value={task.dueDate} className="text-xs" />
 			</div>
 			<div className="flex items-center gap-x-1.5">
