+++ src/features/workspaces/components/create-workspace-form.tsx
@@ -115,7 +115,7 @@ export const CreateWorkspaceForm = ({ onCancel }: CreateWorkspaceFormProps) => {
 											) : (
 												<Avatar className="size-[72px]">
 													<AvatarFallback>
-														<ImageIcon className="size-[36px] text-neutral-400" />
+														<ImageIcon className="size-[36px] text-muted-foreground" />
 													</AvatarFallback>
 												</Avatar>
 											)}
