+++ src/features/workspaces/components/edit-workspace-form.tsx
@@ -190,7 +190,7 @@ export const EditWorkspaceForm = ({
 												) : (
 													<Avatar className="size-[72px]">
 														<AvatarFallback>
-															<ImageIcon className="size-[36px] text-neutral-400" />
+															<ImageIcon className="size-[36px] text-muted-foreground" />
 														</AvatarFallback>
 													</Avatar>
 												)}
